formating

+++ src/Crowdfunding.sol
@@ -7,33 +7,20 @@ contract Crowdfunding is ReentrancyGuard {
     struct Campaign {
         address creator;
         string name;
-        uint goal;
-        uint deadline;
-        uint totalContributed;
+        uint256 goal;
+        uint256 deadline;
+        uint256 totalContributed;
         bool isWithdrawn;
-        mapping(address => uint) contributions;
+        mapping(address => uint256) contributions;
     }
 
-    uint public campaignCount;
-    mapping(uint => Campaign) public campaigns;
+    uint256 public campaignCount;
+    mapping(uint256 => Campaign) public campaigns;
 
-    event CampaignCreated(
-        uint campaignId,
-        string name,
-        uint goal,
-        uint deadline
-    );
-    event ContributionReceived(
-        uint campaignId,
-        address indexed contributor,
-        uint amount
-    );
-    event FundsWithdrawn(uint campaignId);
-    event RefundIssued(
-        uint campaignId,
-        address indexed contributor,
-        uint amount
-    );
+    event CampaignCreated(uint256 campaignId, string name, uint256 goal, uint256 deadline);
+    event ContributionReceived(uint256 campaignId, address indexed contributor, uint256 amount);
+    event FundsWithdrawn(uint256 campaignId);
+    event RefundIssued(uint256 campaignId, address indexed contributor, uint256 amount);
 
     // Errors
     error NameCannotBeEmpty();
@@ -50,16 +37,12 @@ contract Crowdfunding is ReentrancyGuard {
     error RefundFailed();
     error TransferFailed();
 
-    function createCampaign(
-        string calldata _name,
-        uint _goal,
-        uint _duration
-    ) external {
+    function createCampaign(string calldata _name, uint256 _goal, uint256 _duration) external {
         if (bytes(_name).length == 0) revert NameCannotBeEmpty();
         if (_goal == 0) revert GoalMustBeGreaterThanZero();
         if (_duration == 0) revert DurationMustBeGreaterThanZero();
 
-        uint deadline = block.timestamp + _duration;
+        uint256 deadline = block.timestamp + _duration;
 
         Campaign storage newCampaign = campaigns[campaignCount];
         newCampaign.creator = msg.sender;
@@ -71,7 +54,7 @@ contract Crowdfunding is ReentrancyGuard {
         campaignCount++;
     }
 
-    function contribute(uint _campaignId) external payable {
+    function contribute(uint256 _campaignId) external payable {
         Campaign storage campaign = campaigns[_campaignId];
         if (_campaignId >= campaignCount) revert CampaignDoesNotExist();
         if (block.timestamp >= campaign.deadline) revert CampaignExpired();
@@ -83,7 +66,7 @@ contract Crowdfunding is ReentrancyGuard {
         emit ContributionReceived(_campaignId, msg.sender, msg.value);
     }
 
-    function withdrawFunds(uint _campaignId) external {
+    function withdrawFunds(uint256 _campaignId) external {
         Campaign storage campaign = campaigns[_campaignId];
         if (_campaignId >= campaignCount) revert CampaignDoesNotExist();
         if (msg.sender != campaign.creator) revert OnlyCreatorCanWithdraw();
@@ -92,46 +75,41 @@ contract Crowdfunding is ReentrancyGuard {
         if (campaign.isWithdrawn) revert FundsAlreadyWithdrawn();
 
         campaign.isWithdrawn = true;
-        uint amount = campaign.totalContributed;
+        uint256 amount = campaign.totalContributed;
 
         emit FundsWithdrawn(_campaignId);
-        (bool success, ) = campaign.creator.call{value: amount}("");
+        (bool success,) = campaign.creator.call{value: amount}("");
         if (!success) revert TransferFailed();
     }
 
-    function refund(uint _campaignId) external {
+    function refund(uint256 _campaignId) external {
         Campaign storage campaign = campaigns[_campaignId];
         if (_campaignId >= campaignCount) revert CampaignDoesNotExist();
         if (block.timestamp < campaign.deadline) revert CampaignOngoing();
         if (campaign.totalContributed >= campaign.goal) revert GoalNotMet();
-        uint amount = campaign.contributions[msg.sender];
+        uint256 amount = campaign.contributions[msg.sender];
         if (amount == 0) revert NoContributionToRefund();
 
         campaign.contributions[msg.sender] = 0;
 
         emit RefundIssued(_campaignId, msg.sender, amount);
-        (bool success, ) = msg.sender.call{value: amount}("");
+        (bool success,) = msg.sender.call{value: amount}("");
         if (!success) revert RefundFailed();
     }
 
-    function getContribution(
-        uint _campaignId,
-        address _contributor
-    ) external view returns (uint) {
+    function getContribution(uint256 _campaignId, address _contributor) external view returns (uint256) {
         return campaigns[_campaignId].contributions[_contributor];
     }
 
-    function getCampaign(
-        uint _campaignId
-    )
+    function getCampaign(uint256 _campaignId)
         external
         view
         returns (
             address creator,
             string memory name,
-            uint goal,
-            uint deadline,
-            uint totalContributed,
+            uint256 goal,
+            uint256 deadline,
+            uint256 totalContributed,
             bool isWithdrawn
         )
     {
